refactor: remove pagination folder
