--- src/app/(core)/teachers/search/pagination/usePagination.ts
@@ -1,5 +1,0 @@
-export { default } from "@/components/SharedComponents/Pagination/usePagination";
-export type {
-  PaginationProps,
-  UsePaginationProps,
-} from "@/components/SharedComponents/Pagination/usePagination";
